refactored the code to pay some technical debt
